requirements.txt
fastapi>=0.128.0
uvicorn>=0.30.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
sse-starlette>=2.0.0
python-dotenv>=1.0.0
